Use proper requests/aiohttp sesssions
After consideration, I decided to properly use
requests and aiohttp sessions even though they
complexify the examples. For aiohttp the direct
request calls are deprecated, and for requests the
only time you should not use sessions is if you
don't care about performance - this isn't the case
in this talk.

This also allows for some discussion of the fact
that requests sessions are (AFAIK) not considered
threadsafe.